perf: use float32 for CSR data array to halve memory
The data array for work_referenced_works (3B edges) at float64 is 24 GB.
float32 is sufficient for unweighted (all-ones) adjacency matrices and
cuts this to 12 GB, reducing peak CSR construction memory from ~50 GB
to ~38 GB.

+++ sync/build_csr.py
@@ -350,7 +350,7 @@ def _build_csr_duckdb(
         indices = np.concatenate(indices_chunks)
         del indices_chunks
 
-        data = np.ones(n_edges, dtype=np.float64)
+        data = np.ones(n_edges, dtype=np.float32)
         csr = sparse.csr_matrix(
             (data, indices, indptr),
             shape=(n_nodes, n_nodes),
